Moving deduceTransactionBodyForFundToken
